+++ CHANGELOG.md
@@ -5,8 +5,9 @@
     * `PreviewCustomizationChoice` allows you to view a customization choice using a ChrCustomizationOptionID and a ChrCustomizationChoiceID
     * `SetDressState` allows you to manually toggle your character's dress state using 1 or 0
     * If the console is not enabled, you can enable it with `/console enable` or by launching your game with the `-console` flag
-* Added a new slash command `/dm appearanceinfo` to return some transmog data from an AppearanceID
-* Added a command `/dm tryonitem` to try on an item by itemID. Note that item IDs can have multiple appearances, so this might not be 100% what you're expecting.
+* Added a new slash command `/dm appearanceinfo` to return some transmog data from an itemAppearanceID
+* Added a command `/dm appearancemods` to return all itemModifiedAppearanceIDs for a given itemAppearanceID
+* Added a command `/dm tryonitem` to try on an item by itemID. Note that item IDs can have multiple appearances, so this might not be 100% what you're expecting
 
 ##### Fixed
 * Fixed an infinite loop in the `/dm quest` command